Use linked event number for all stream transient subscriptions
