+++ lib/extreme.ex
@@ -12,6 +12,7 @@ defmodule Commanded.EventStore.Adapters.Extreme do
 
   alias Commanded.EventStore.Adapters.Extreme.{
     Config,
+    Mapper,
     PubSub,
     Subscription,
     SubscriptionsSupervisor
@@ -264,7 +265,7 @@ defmodule Commanded.EventStore.Adapters.Extreme do
         read_events = read_events ++ events
 
         if end_of_stream? || length(read_events) == count do
-          recorded_events = Enum.map(read_events, &to_recorded_event/1)
+          recorded_events = Enum.map(read_events, &Mapper.to_recorded_event/1)
 
           {:ok, recorded_events, end_of_stream?}
         else
@@ -286,68 +287,6 @@ defmodule Commanded.EventStore.Adapters.Extreme do
     end
   end
 
-  def to_recorded_event(%ExMsg.ResolvedIndexedEvent{event: event, link: nil}),
-    do: to_recorded_event(event, event.event_number + 1)
-
-  def to_recorded_event(%ExMsg.ResolvedIndexedEvent{event: event, link: link}),
-    do: to_recorded_event(event, link.event_number + 1)
-
-  def to_recorded_event(%ExMsg.ResolvedEvent{event: event}),
-    do: to_recorded_event(event, event.event_number + 1)
-
-  def to_recorded_event(%ExMsg.EventRecord{} = event),
-    do: to_recorded_event(event, event.event_number + 1)
-
-  def to_recorded_event(%ExMsg.EventRecord{} = event, event_number) do
-    %ExMsg.EventRecord{
-      event_id: event_id,
-      event_type: event_type,
-      created_epoch: created_epoch,
-      data: data,
-      metadata: metadata
-    } = event
-
-    data = deserialize(data, type: event_type)
-
-    metadata =
-      case metadata do
-        none when none in [nil, ""] -> %{}
-        metadata -> deserialize(metadata)
-      end
-
-    {causation_id, metadata} = Map.pop(metadata, "$causationId")
-    {correlation_id, metadata} = Map.pop(metadata, "$correlationId")
-
-    %RecordedEvent{
-      event_id: UUID.binary_to_string!(event_id),
-      event_number: event_number,
-      stream_id: to_stream_id(event),
-      stream_version: event_number,
-      causation_id: causation_id,
-      correlation_id: correlation_id,
-      event_type: event_type,
-      data: data,
-      metadata: metadata,
-      created_at: to_naive_date_time(created_epoch)
-    }
-  end
-
-  defp to_stream_id(%ExMsg.EventRecord{event_stream_id: event_stream_id}) do
-    event_stream_id
-    |> String.split("-")
-    |> Enum.drop(1)
-    |> Enum.join("-")
-  end
-
-  defp to_naive_date_time(millis_since_epoch) do
-    secs_since_epoch = round(Float.floor(millis_since_epoch / 1000))
-    millis = :erlang.rem(millis_since_epoch, 1000)
-    epoch_secs = :calendar.datetime_to_gregorian_seconds({{1970, 1, 1}, {0, 0, 0}})
-    erl_date = :calendar.gregorian_seconds_to_datetime(epoch_secs + secs_since_epoch)
-
-    NaiveDateTime.from_erl!(erl_date, {millis * 1000, 3})
-  end
-
   defp read_events(stream, from_event_number, max_count, direction) do
     msg_type =
       if :forward == direction, do: ExMsg.ReadStreamEvents, else: ExMsg.ReadStreamEventsBackward
@@ -360,10 +299,6 @@ defmodule Commanded.EventStore.Adapters.Extreme do
       require_master: false
     )
   end
-
-  defp serialize(data), do: Config.serializer().serialize(data)
-
-  defp deserialize(data, opts \\ []), do: Config.serializer().deserialize(data, opts)
 
   defp add_causation_id(metadata, causation_id),
     do: add_to_metadata(metadata, "$causationId", causation_id)
@@ -405,6 +340,8 @@ defmodule Commanded.EventStore.Adapters.Extreme do
       require_master: false
     )
   end
+
+  def serialize(data), do: Config.serializer().serialize(data)
 
   # Event store supports the following special values for expected version:
   #
